o/snapstate: account for deleted revs when undoing install
The undoLinkSnap handlers puts the candidate revision back in its
previous position, if the install fails. However, the install may
fail after the discard-snap tasks run which means that the old
candidate index may be outdated which could result in the wrong
Sequence being remade or even a panic. This change accounts for the
number of discarded revisions when undoing the install.