packaging: make postrm script robust against `rm` failures
We do not want to risk failing the whole script (and therefore leaving
the user's system in a dirty state) if removing the snap directories
fails.
This has happened in real life (see the bug report linked below) and,
although the mechanisms to reproduce the issue are not clear, this fixes
the symptoms.

Fixes: https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1950634