gadget: remove ID from DiskStructureDeviceTraits
This was meant to be something like /dev/disk/by-id, but actually on some
systems there are multiple values for the same partition/disk that qualify as
"ID's", i.e.

$ for p in $(udevadm info --query property /dev/nvme0n1p1 | grep -Po 'DEVLINKS=\K.*'); do if [ $(dirname $p) = /dev/disk/by-id ]; then echo $p; fi; done
/dev/disk/by-id/nvme-Samsung_SSD_970_EVO_Plus_1TB_S4H7NF1M476095R-part1
/dev/disk/by-id/nvme-eui.0025222491c334d2-part1

So rather than try and change this to an array and then have to do set
operations to check if there is an overlap or if two sets are equal to each
other, just remove it as we have plenty of other properties to measure about
disks and partitions that should be sufficient.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>