gadget: allow EnsureLayoutCompatibility to ensure disk has all laid out structs
This is important for validating that a fully laid out gadget.yaml and a disk
which should already have all partitions created on it (i.e. after install
mode on UC20+) are compatible. Previously, without this it was possible for
a disk to be compatible even when non-creatable structures in the gadget.yaml
were not found on disk. This will be mainly important for gadget asset updates.

Also adjust unit tests for reality, since on a real system (both DOS and GPT
schemas), there is no device node for the mbr structure, so it will not appear
as /dev/node1, it will simply be absent and the code and tests should account
for this.

Due to the absence of the MBR structure in ondisk volumes, we also need to make
an exception for the mbr structure when validating that all structures in the 
YAML are present ondisk too. 

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>