bootloader/assets: add fallback entry to grub.cfg
Add fallback entry to grub configuration, so in updates we boot the
previous kernel if the new one fails to be loaded by grub or the
systemd stub. This is needed when failures happen before exitting UEFI
boot services: otherwise we get stuck in a grub menu and a manual
reboot would be needed. The case when the kernel itself fails should
be handled by a hw watchdog.

Also, remove the test for the kernel file so the fallback mechanism
also works when the try-kernel.efi link has not been
created (previously we would get stuck in grub prompt if that
happened).