gadget: leave TODOs about handling multiple potential matches
For DOS, we probably should actually do something sensible with the multiple
errors, as there is no way to know if a structure/partition "should match"
since DOS partitions don't have partition labels the same way that GPT
partitions do.

But for GPT disks, we should bail early since if the name matches we should not
be able to have multiple partitions with the same name on a single volume.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>