bootloader/assets: on failure, reboot instead of chainloading
If we chainload in the case of a failed kernel update, we will not be
able to unseal keys from the TMP, as the measurements in the PCR
registers store all the history since we booted so we will get there
unexpected values in the fail case. Instead, reboot so we clean the
PCR registers and we can start the old kernel.