cmd/snap-confine: do not include libglvnd libraries from the host system
* cmd/snap-confine: do not include libglvnd libraries from the host system

* tests: we no longer symlink libGLX.so

* cmd/snap-confine: include glvnd globs for old "base: core" snaps

Ubuntu 16.04 did not include the glvnd driver multiplexing libraries,
and the Mesa version of libGL will not function with the Nvidia X
drivers.

While the glvnd drivers may not be compatible with the libraries in the
snap's sandbox (e.g. we know that Ubuntu 21.10's libEGL uses new glibc
symbols), it is better than the nothing working. In particular, X11
based OpenGL apps will function, which will cover the majority of these
old snaps.

* tests: adjust opengl-nvidia test to use its own test snap.

Also update to run on Ubuntu 20.04, and drop 14.04. Still todo: test
against the core20 test snap.

* cmd/libsnap-confine-private: add a unit test for sc_cleanup_shallow_strv

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* tests: don't specify base for gl-core16 test snap

* tests: show that host system GLVND libraries are not exposed to snaps using newer bases

* tests: do not run the gl-core20 part of the test on i386

* tests: show that the nvidia backend drivers are still available on core20

* tests: adjust spread test

* tests: add back missing canary file

* cmd/snap-confine: only create the globs array if NVIDIA_BIARCH||NVIDIA_MULTIARCH

Co-authored-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>