cmd/snap: print a placeholder for version of broken snaps
When producing the output of `snap list`, snaps that are broken (i.e. not
mounted) are printed without their version. This can be confusing when the
output is further processes by shell tools such as sed/awk/cut. Tweak the code
to always include a placeholder, even if the version is empty.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>