tests/main/cgroup-tracking-failure: Make it pass when run alone
This test does not behave the same if it was run alone or after other
tests like `tests/main/cgroup-tracking` or
`tests/main/cgroup-tracking`. So we need to simulate the other tests
bad cleanup to make it succeed alone.

It is not possible to properly cleanup the root session without having
all ssh connection restarted.