gadget: allow gadget struct with unspecified filesystem to match part with fs
This is primarily to enable devices where there are extra partitions on the
disks that are not used with Linux and are instead just treated as binary blobs
for other co-processors, etc. and the binary blob just so happens to be an
actual image of a filesystem, which leads to discovering the on-disk device as
having a filesystem, but the gadget saying there is no filesystem.

Before this, we would fail to match these two structures and required instead
specifying the filesystem in the gadget.yaml, but that would mean not
specifying the content of the partition in the gadget.yaml with an image which
is awkward and difficult.

Also fundamentally re-organize the lambda function to check a gadget and
on-disk structure to be more readable and straight-forward.

Finally add a unit test for this case.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>