tests: revert ef1e9abb71c (workaround for systemd 245->246)
The code ef1e9abb71c in prepare-restore.sh was added to workaround
a systemd upgrade from v245 to v246 in groovy (20.10) that only
happend during the 20.10 development cycle. We no longer need
this workaround and it actually caused harm because using
`install_pkg_dependencies` inside an `if` prevented dependency
installs.