gadget: resolve index ambiguity between OnDiskStructure and LaidOutStructure
We had an issue previously, where we have both LaidOutStructure.Index and then
also (because LaidOutStructureIndex is embedded into OnDiskStructure)
OnDiskStructure.Index, but the meaning/usage of these two indices are subtly
different. The former is 0-referenced, and refers to the order in which all the
structures are mentioned in the abstract gadget.yaml, while the latter is
1-referenced and refers only to all structures which physically exist and are
measurable from the partition table (this excludes specifically MBR and other
type: bare structures).

To resolve this ambiguity and easy confusion, introduce a specific
OnDiskStructure.DiskIndex as well as rename the LaidOutStructure.Index to 
YamlIndex to make clear what the indices are referring to. Additionally, when
creating an OnDiskStructure to prevent possible confusion, use a new constant
UnknownLaidOutStructureIndex to make even more clear that OnDiskStructure's 
should not have their LaidOutStructure.YamlIndex used and should instead be 
using the DiskIndex.

Also split out the creation of an OnDiskStructure from a disks.Partition into
a new helper function since this will need to be done in other places too.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>