gadget: introduce OnDiskStructure.StructureIndex, use where applicable
We had an issue previously, where we have both LaidOutStructure.Index and then
also (because LaidOutStructureIndex is embedded into OnDiskStructure)
OnDiskStructure.Index, but the meaning/usage of these two indices are subtly
different. The former is 0-referenced, and refers to the order in which all the
structures are mentioned in the abstract gadget.yaml, while the latter is
1-referenced and refers only to all structures which physically exist and are
measurable from the partition table (this excludes specifically MBR and other
type: bare structures).

To resolve this ambiguity and easy confusion, introduce a specific
OnDiskStructure.StructureIndex as well as a constant
UnknownLaidOutStructureIndex, to be used when creating an OnDiskStructure since
LaidOutStructure is embedded inside and thus inherits the Index member
directly.

Also split out the creation of an OnDiskStructure from a disks.Partition into
a new helper function since this will need to be done in other places too.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>