cmd/snap: setup tracking cgroup when trying to invoke a service as a user
Normally apps declared to be services are started by systemd, and thus end up in
a service specific cgroup. However, it is possible for a user to invoke `snap
run foo.bar` (where foo.bar is a service) and thus run it themselves. In such
scenario, the service will be in the user's cgroup (typically the session
scope). On a cgroup v2 system, should the service then get access to devices,
and attempt to modify device filtering on the cgorup could break the session,
thus snap-confine stops and dies before modifying the cgroup. This was not a
problem with cgroup v1 as the device controller was separate and used a disjoint
hierarchy which could be modified freely.

Attempt to remedy this, by establishing a tracking cgroup also when an app being
a service is not in a systemd created group, in which case snap will attempt to
create a transient scope for it, like it does does all other apps.

Fixes: https://bugs.launchpad.net/snapd/+bug/1956917
See: https://forum.snapcraft.io/t/session-101-scope-is-not-a-snap-cgroup/28193

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>