tests: fix `tests/core/create-user` on testflinger pi3
The `create-user` test was failing on the testflinger external
pi3 devices. The reason is that the test only looked at
`$MANAGED_DEVICE` which is set in spread.yaml just based on
if the device is external or not. However not all external
devices are already managed (the testflinger pi3 is not).

So this commit fixes this by simply checking directly for
`snap managed` instead.