many: add helper function for type conversions
Previously, the assignment made via the Attrer.Attr() method were only
happening if the destination variable had exactly the same type as the
input value. This is quite limiting, because in our use-cases the input
value comes from the YAML and types are always wrapped in golang
interfaces; for example, a list of strings would not be represented in a
`[]string` but rather in a `[]interface{}`. This means that all complex
types could not be retrieved.

For the time being, this function only adds support for converting
arrays. Map conversions will be added once we have a need for them.