snap-bootstrap: Cleanup dependencies in systemd mounts
To be able to survive switch root, mounts need to be "wanted" directly
or indirectly by `initrd-switch-root.target`. `initrd-fs.target` is
the proper target and enough. Specifying other targets is superfluous.

The mounts do not need to be "required". "Requires" has a stricter
semantic and potentially makes systems less robust. It is recommended
by documentation that we use "Wants" where "Requires" is not needed.

`initrd.target` and `local-fs.target` are not part
`initrd-switch-root.target` and are normally taken down during the
switch of root.

Default dependencies are not useful in non ephemeral mounts.
Specifically `Before=local-fs.target`. The dependency we need is
`Before=initrd-fs.target`.