gadget/install/partition.go: use device rescan trick only when gadget says to
This rescan file does not seem to properly work in either the GCE cloud VM's we
use for spread testing, nor on ordinary QEMU VM's with IDE storage devices, so
we only want to do it for the very specific device that absolutely needs it
(because executing partx will break kernel data structures for the disks and
this kernel bug will not be fixed for various reasons).

So instead only execute the rescan trick if the gadget instructs us to by
placing a file named "force-partition-table-reload-via-device-rescan.txt" in
the root of the gadget.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>