mkversion: check that snapd is a git source tree before guessing the version
It is possible that when the mkversion is called, the snapd source tree isn't a
git repository, but rather an extracted source tarball.

To make matters worse, the distro packaging files can be kept in a git source
tree. Then during the build from source tarballs within that source tree
mkversion will incorrectly interpret the output of `git describe` which actually
describes the state of the distro packaging repository.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>