tests: add back systemd-timesyncd to newer debian distros
This package was initially added in #10949 to all debian distros in
order to fix the interfaces-timeserver-control test, but was recently
reverted in 29f24fb558209ae492c2a4eba0453feda98e0451 as the package was
not present in debian-10.

So let's try to add it only to debian sid and debian 11.