osutil: refactoring the code exporting mocking APIs to other packages
* osutil: Refactoring the code exporting mocking APIs to other packages

File 'mockable.go' is used to mock certain data obtained by 'osutil'
from the system (e.g. 'mountinfo'). The file also contained some
variables that belong to production code itself. A cleanup
refactoring was applied to this file and 'mountinfo_linux.go':
- Package variables moved from 'mockable.go' to logic that owns it
- Test-related logic moved from 'mountinfo_linux.go' to 'mockable.go'
- United and refactored logic designed to mock 'mountinfo'

Signed-off-by: Arseniy Aharonov <arseniy.aharonov@canonical.com>

* osutil: Fixes and cleanups after the code review

Signed-off-by: Arseniy Aharonov <arseniy.aharonov@canonical.com>

* osutil: make mustNotBeTestBinary function non-exportable

To avoid unwonted usage of this method in different places in
the code it was requested to make this function local to package
osutil only.

Signed-off-by: Arseniy Aharonov <arseniy.aharonov@canonical.com>

* osutil: refactoring mountinfo code following PR review

My original refactoring of the mountinfo logic introduced
an inconsistency into the code, so that os-agnostic code
contain Linux-related logic. This inconsistencey was
eliminated.

Signed-off-by: Arseniy Aharonov <arseniy.aharonov@canonical.com>