o/snapstate: write state when undoing migration
Write state and sequence file after undoing the migration, if
these exist. This must be done after the undo procedure because
only then do we know what the migration's final state is. Also,
take care to re-write what was written during the change (e.g., if
the sequence file was written but not the state, re-write only the
sequence file with the undone migration status).
Add comments explaining why state is written where it is.