tests: fix snap-run-gdbserver restore
The test was first installing the `gdb{,server}` packages and then
removing them again. On the 22.04 image these packages are already
installed and removing them again triggers code in restore_suite_each()
that will just reinstall them. If the distro is in flux (like 22.04)
this can even confuse apt so it's best to just avoid the (unneeded)
`apt autoremove`.