spread: disable enforcing mode on CentOS 8 Stream
The default image is configured in SELinux enforcing mode, which isn't quite
compatible with tests that touch random things in the system. Switch to
permissive mode early in the setup. Note, there are tests that verify snapd's
SELinux policy that explicitly enable enforcing mode for the duration of the
test.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>