gadget/install: rm unused support for writing non-filesystem structures
This was never used, since by design, we only create partitions with
filesystems in install mode (and specifically only ubuntu-save, ubuntu-data,
and ubuntu-boot).

Also remove the gadgetRoot parameter since it is not used.

Removing support for this isn't really necessary, but it's unused code that we
likely do not ever have a plan for using.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>