gadget: identify/match encryption parts, include in traits info
Using out of band information about what partitions are expected to be
encrypted on a given disk that we are verifying is compatible with a gadget
layout, validate that encrypted partitions are indeed encrypted on disk.

Also include this information in the disk-mapping.json as this will be useful
when performing gadget asset updates on encrypted devices, as we will not need
to query or otherwise determine which structures are encrypted if we have this
information available from the disk-mapping.json.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>