interfaces, o/ifacestate: generate profile transition rules for devmode snaps
These additional accesses allow a snap that is in devmode confinement to
execute other snaps via `snap run ...` while snap-confine still ends up being
confined with a strict apparmor profile. Previously what would happen is that
snap-confine when executed from a devmode snap would only be in complain mode,
which is the same as the devmode confinement of the calling snap. This is not
ideal because it exposes snap-confine to attacks that may lead to privilege
escalation because unprivileged users can execute commands from inside a
devmode snap.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>