snap-confine: ensure that snap-confine is in `strict` confinement
This commit makes the check in `sc_init_apparmor_support()` stricter
when it checks if snap-confine itself is running confined. Before
it only checked that there is the expected profile but it would
not check if the profile is in strict or complain mode. So a
snap-confine in complain mode was passing as `is_confined`. With
this commit `is_confined` is only true if it's actually in strict
confinement. This is another hardening step to make attacking
snap-confine more difficult.