gadget/gadget_test.go: fix variable type
This type got moved to gadgettest in one of the prior PRs, but the checks for
the PR where this specific test was introduced had it's checks run before that
PR was merged so it was green and thus mergable and didn't show this error.

Thanks to Maciej for finding this out so quickly.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>