o/snapstate/snapmgr: remove inactive, vulnerable snapd, core snap revisions
Only do this when the active revision of snapd or core is fixed however.

This resolves an issue with the recent fixes for the security vulnerability
wherein old versions of the vulnerable snap-confine were left behind on disk,
where they could be abused to continue executing the vulnerability that has
since been fixed. Eventually we will roll out a fix where old versions of snapd
and core are mounted nosuid so that snap-confine inside them cannot be abused
this way, but until that happens do the radical thing of removing vulnerable
revisions automatically when the fixed revisions are successfully installed.

Whether or not a revision is fixed is indicated by the new snapd/info flag
SNAP_CONFINE_CVE_2021_44731.

Also update unit tests to mock this file since we now need to read it in many
snapstate and overlord tests.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>