osutil/disks: Calculate the last usable LBA instead of reading it
Installation of Ubuntu Core is usually done by writing an image onto a
medium. The size of the medium is usually bigger than the image.  It
is expected that the partition `ubuntu-data` will be sized to fill up
that medium.

Because GPT has a backup header and table at the end of the disk, it
is not possible to create a partition until the end of this disk.
For this reason the entry "last usable LBA" from the header
tells us until where we can create partitions.

For the case of Ubuntu Core, since the size of the medium is typically
bigger than the image, the backup header and table will not be at the
end of the disk when reading it. It will be fixed later when updating
the table.

For that reason, the "last usable LBA" from the image will not
be the correct one for the disk.

`sfdisk --json` used to report a calculated "last usable LBA" instead
of the effective value stored in the header. But it has been changed
since 387ac2771f1b16af00c417b226c697d08d1301fe (2.36), and now it
reports the value from the header.

For that reason we have to calculate the expected "last usable LBA"
that will be used. This is calculated from the size of the table
and the size of the disk.

While it is possible to also to `echo write | sfdisk /dev/disk`,
to fix before `sfdisk --json`, we would require to wait for
udev to treat all events.