gadget/update.go: fix old case of single-volume gadgets without system-boot
Historically we were not as strict with single-volume gadget validation and
thus allowed installation of gadgets which did not have any reference to a
system-boot role, and we need to continue to support these, so we do indeed
need to have a branch for the single volume case (I had inadvertently
refactored that away in the previous code without realizing the error).

Also add unit tests for this specific case.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>