snap-bootstrap: Partially revert simplifications of mount dependencies
In a previous change, we removed a lot of useless dependencies in
mounts generated by `snap-bootstrap initramfs-mounts`. But because of
some cleanup was done also on core22, we get an issue with a missing
dependency from `local-fs.target`. Mounts get unmounted in the main
boot of core22 (not core20) because they are not part of any active
target.