kernel/fde: add PartitionName to various structs
We need to specify the PartitionName for ICE decryption since it is currently
based on partition name as well as the mapper device provided in Device.

Also allow mocking the actual call to runFDEDeviceUnlockCommand from other 
packages without all the complexity of creating hook scripts and such, and add
some more debug logging.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>