asserts,interfaces/policy: slot-snap-id allow-installation constraints
The use case for this is that we have now the unsupported situation of
an interface that is superprivileged but still wants to have special
system snap slots.

Because it needs to cover also the --dangerous case the
base-declaration slot side slot-snap-type constraint cannot express
this, it needs to allow both core and app. To restrict this as needed
we then allow to use slot-snap-id which can list the well-known system
snap ids.

Notice that this kind of constraint makes sense only in the
base-declaration.  In a snap-declaration the snap-id is fixed and
implied. We also do not want to put interface rules in the actual
snap-declarations of the system snaps.

Although there's no clear use case, as the system snaps don't have
plugs, we support also plug side allow-installation plug-snap-id for
symmetry.