debug-tools/snap-debug-info.sh: add debug script
This is the first iteration of the debug script for debugging customer issues,
with a collection of easily gathered and commonly requested information that is
helpful in debugging issues on Ubuntu Core.

Eventually this will become a snap command and a REST API endpoint, but for now
a script will have to do.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>