o/snapshotstate: add an extra task for performing the cleanup
This ensures that the cleanup will be performed immediately, and not
deferred to a later time. The cleanup-restore-snapshot task however will
only be executed if all the restore tasks succeed, so we still have to
register the cleanup handler for failure situations.
The cleanup is idempotent, so this shouldn't cause harm.