debug-tools: add script to help debugging GCE instances which fail to boot
When debugging GCE spread runs where the VM fails to boot properly, you need to
look at the serial console, however doing so with the gcloud command line tool
is a bit finnicky.

This is a bash script which essentially will collect and follow all of the
output on the serial console from a GCE instance.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>