packaging: use latest go to build spread
The most recent change in `go get` with go1.18 did not longer
install binaries. However we used this feature in our autopkgtest
script. As a workaround there was a commit to build spread with
go 1.17 to workaround this change.

This commit moves us back to latest go and instead of `go get`
we now use `go install`.