snap: implement "star" developers
This commit adds support for the new account `starred` validation
value. It displays a unicode star "✪" when possible, otherwise
a ascii "*". Note that this also means that validated developers
get a "**" appended now. While this is an incompatible change
it is what was agreeded in the star developers spec because it
is the most natural way to present these values in ascii.