tests/main/user-session-env: for for opensuse
The test is currently failing for opensuse-15.3:

    + for user in test "$TEST_ZSH_USER" "$TEST_FISH_USER"
    + echo 'checking test-fish'
    checking test-fish
    + '[' -e test-fish-session-env ']'
    + case "$user:$SPREAD_SYSTEM" in
    + MATCH 'XDG_DATA_DIRS=.[:]?/var/lib/snapd/desktop[:]?.'
    + MATCH 'PATH=.[:]?/snap/bin[:]?.'
    grep error: pattern not found, got:
    HOME=/home/test-fish
    LANG=en_US.UTF-8
    LOGNAME=test-fish
    PATH=/usr/local/bin:/bin:/usr/bin
    ENV_ROOTPATH	/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin:/snap/bin
    PWD=/home/test-fish
    SHELL=/usr/bin/fish
    SHLVL=1
    USER=test-fish
    XDG_DATA_DIRS=/usr/local/share:/usr/share:/var/lib/snapd/desktop

It seems openSUSE 15.3 is affected by the same issue as Tumbleweed, so
let's treat them in the same way. It's possible that the issue is not
caused by fish, but by some other system package (since the version of
fish in the two openSUSE distros we support is very different).