snapstate: do not auto-migrate to ~/Snap for core22 just yet
During the roadmap sprint in Copenhagen we had a discussion about
the behavior of the ~/Snap folder migration. Unfortunately there
seems to have been some misunderstandings in the previous plan
and during the discussion it became apparent that is considered
to risky to migrate all the snaps with `base: core22`.

Instead we will most likely create an experimental option to
opt-in into the ~/Snap folder for selected snaps to get feedback
and gather data what snaps behave well and which are problematic.

Based on these results we will then decide if:
a) we do the migration at a future point automatically for all snaps
b) if allow snaps to opt-in into the migration via a flag in snap.yaml
c) migrate with core24

This is an initial commit to remove the current automatic migration
for core22 based snaps. We need to act quickly here to ensure
as few snaps as possible have migrated already (but there really
should be none because snapcraft does not allow building
`base: core22` snaps yet in their stable channel).