cmd/snap-confine: improve comment
The comment was mentioning the same function twice. Looking at the git
history, it appears that the reason for the comment was this remark
raised during the review process:

  https://github.com/snapcore/snapd/pull/4312#discussion_r153789828

So, besides fixing the function name, expand the comment a bit to
clarify exactly why the functions might need to be updated.